third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/cssom/cssstylesheet-insertrule-blocked-expected.txt

CONSOLE ERROR: Refused to evaluate a string as CSS because 'unsafe-eval' is not an allowed source of style in the following Content Security Policy directive: "style-src 'unsafe-inline' 'self'".

CSSStyleSheet::insertRule blocked if 'unsafe-eval' is not an allowed source of style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.styleSheets[0].insertRule('x { display: none; }', 0); threw exception SecurityError: Failed to execute 'insertRule' on 'CSSStyleSheet': Refused to evaluate a string as CSS because 'unsafe-eval' is not an allowed source of style in the following Content Security Policy directive: "style-src 'unsafe-inline' 'self'".
.
PASS successfullyParsed is true

TEST COMPLETE
